Deepsort to sdk

+++ gen2-deepsort-tracking/deep_sort_realtime/deep_sort/detection.py
@@ -32,7 +32,7 @@ class Detection(object):
 
     def __init__(self, ltwh, confidence, feature, class_name=None, others=None, id=-1):
         # def __init__(self, ltwh, feature):
-        self.ltwh = np.asarray(ltwh, dtype=np.float)
+        self.ltwh = np.asarray(ltwh, dtype=np.float32)
         self.confidence = float(confidence)
         self.feature = np.asarray(feature, dtype=np.float32)
         self.class_name = class_name
